pyproject.toml
[project]
name = "mtnchatbot"
version = "0.1.0"
description = "MTN RAG Chatbot - Upload PDFs and ask questions about their content"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "groq>=0.31.1",
    "pypdf2>=3.0.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "python-dotenv>=1.0.0",
]
